+++ src/docuware/errors.py
@@ -9,10 +9,12 @@ class DocuwareClientException(Exception):
         *args,
         url: Optional[str] = None,
         status_code: Optional[int] = None,
+        server_message: Optional[str] = None,
     ):
         super().__init__(*args)
         self.url = url
         self.status_code = status_code
+        self.server_message = server_message
 
     def __str__(self) -> str:
         if self.url: